refactor(confusables): extract the TR39 signature to one shared definition
The visible mixed-script confusable table and its decision rule had one home,
lint_confusables.ts, and the runtime encoding scanner needs the same signature.
Copying it would guarantee drift: a codepoint added to one and not the other
means the authoring gate and the runtime floor disagree about what a homoglyph
is.

Extracted to _lib/confusables.ts; _classify_token becomes a re-export of the
shared classifyToken, so the linter's published surface and its tests are
unchanged.

Extraction rather than importing lint_confusables directly, because the
dependency direction matters: the shared module imports NOTHING, so the runtime
retrieval path does not pull the authoring-time lint framework (fs,
security_lint) into a request.

Behaviour-preserving: lint_confusables still reports clean over the corpus and
its 13 tests pass unchanged.
